+++ examples/test_quickstart.py
@@ -1,44 +1,71 @@
 import os
 import io
+import ast
+import sys
+import json
+import tempfile
 import unittest
 import contextlib
+import subprocess
 import unittest.mock
 import importlib.util
 
+from dffml.util.os import chdir
+
+
+def sh_filepath(filename):
+    return os.path.join(os.path.dirname(__file__), "quickstart", filename)
+
+
+@contextlib.contextmanager
+def directory_with_csv_files():
+    with tempfile.TemporaryDirectory() as tempdir:
+        with chdir(tempdir):
+            subprocess.check_output(["sh", sh_filepath("train_data.sh")])
+            subprocess.check_output(["sh", sh_filepath("test_data.sh")])
+            subprocess.check_output(["sh", sh_filepath("predict_data.sh")])
+            yield tempdir
+
 
 class TestQuickstart(unittest.TestCase):
-    def run_quickstart(self):
-        """
-        The quickstart code doesn't have a if __name__ == "__main__" block, so
-        importing it will cause it to run.
-        """
-        spec = importlib.util.spec_from_file_location(
-            "quickstart",
-            os.path.join(os.path.dirname(__file__), "quickstart.py"),
-        )
-        common = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(common)
-
-    def test_quickstart(self):
+    def python_test(self, filename):
+        # Path to target file
+        filepath = os.path.join(os.path.dirname(__file__), filename)
         # Capture output
-        stdout = io.StringIO()
-        with contextlib.redirect_stdout(stdout):
-            self.run_quickstart()
-        lines = stdout.getvalue().split("\n")
-        # Check output of Accuracy
-        self.assertIn("Accuracy ", lines[0])
-        self.assertIn(".", lines[0])
-        # Check that repos have predictions
-        rows = list(filter(lambda line: "value:" in line, lines))
-        rows = list(
-            map(
-                lambda line: line.replace("value:", "")
-                .replace("\t", "")
-                .split(",")[0],
-                rows,
+        stdout = subprocess.check_output([sys.executable, filepath])
+        lines = stdout.decode().split("\n")
+        # Check the Accuracy
+        self.assertIn("Accuracy: 1.0", lines[0])
+        # Check the salary
+        self.assertEqual(ast.literal_eval(lines[1])["Salary"], 70)
+        self.assertEqual(ast.literal_eval(lines[2])["Salary"], 80)
+
+    def test_python(self):
+        self.python_test("quickstart.py")
+
+    def test_python_async(self):
+        self.python_test("quickstart_async.py")
+
+    def test_python_filenames(self):
+        with directory_with_csv_files() as tempdir:
+            self.python_test("quickstart_filenames.py")
+
+    def test_shell(self):
+        with directory_with_csv_files() as tempdir:
+            # Run training
+            subprocess.check_output(["sh", sh_filepath("train.sh")])
+            # Check the Accuracy
+            stdout = subprocess.check_output(
+                ["sh", sh_filepath("accuracy.sh")]
             )
-        )
-        row0, row1 = rows
-        # Ensure predictions are floats
-        float(row0.split()[0])
-        float(row1.split()[0])
+            self.assertEqual(stdout.decode().strip(), "1.0")
+            # Make the prediction
+            stdout = subprocess.check_output(["sh", sh_filepath("predict.sh")])
+            repos = json.loads(stdout.decode())
+            # Check the salary
+            self.assertEqual(
+                int(repos[0]["prediction"]["Salary"]["value"]), 70
+            )
+            self.assertEqual(
+                int(repos[1]["prediction"]["Salary"]["value"]), 80
+            )