+++ CHANGELOG.md
@@ -57,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Subclasses of `BaseConfigurable` will now auto instantiate their respective
   config classes using `kwargs` if the config argument isn't given and keyword
   arguments are.
+- The quickstart documentation was improved as well as the structure of docs.
 ### Fixed
 - CONTRIBUTING.md has `-e` in the wrong place in the getting setup section.
 - Since moving to auto `args()` and `config()`, BaseConfigurable no longer